refactor: atualiza shop.js para index.js
