+++ src/pages/on-sale.js
@@ -21,7 +21,7 @@ const OnSale = ({ data }) => {
   return (
     <Layout title="Shop">
       <div className="w-full">
-        <h1 className="text-black font-light text-5xl md:text-6xl mb-12">
+        <h1 className="text-black font-light text-4xl sm:text-5xl md:text-6xl mb-12">
           Plants.<span className="text-green-600">OnSale</span>
         </h1>
         <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-12">
